[vfs] E: Split Descriptor Domain Types
Move descriptor flags, routing, terminal errors, and process-reclaim
result types into dedicated modules. Re-export every type through fd so
existing vfsd and integration consumers retain their current API paths.
