+++ src/libs/vfs/src/descriptor.rs
@@ -11,9 +11,14 @@ mod console_handle;
 mod console_stream;
 mod direct_read_handle;
 mod directory_handle;
+mod fd_flags;
 mod host_fs_handle;
+mod pipe_closure;
+mod process_exit_reclaim;
 mod socket_handle;
+mod tty_error;
 mod vfs_file_handle;
+mod vfs_route;
 mod vfs_stat;
 
 //==================================================================================================
@@ -25,8 +30,13 @@ pub use self::{
     console_stream::ConsoleStream,
     direct_read_handle::DirectReadHandle,
     directory_handle::DirectoryHandle,
+    fd_flags::FdFlags,
     host_fs_handle::HostFsHandle,
+    pipe_closure::PipeClosure,
+    process_exit_reclaim::ProcessExitReclaim,
     socket_handle::SocketHandle,
+    tty_error::TtyError,
     vfs_file_handle::VfsFileHandle,
+    vfs_route::VfsRoute,
     vfs_stat::VfsStat,
 };
